# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "To Do List",
    "description": "To Do LIst REST API",
    "contact": {
      "email": "rmukunda.27@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "http://todobe-env.eba-bcfh3zx8.us-east-2.elasticbeanstalk.com",
  "basePath": "/",
...
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/api/user/create": {
      "post": {
        "tags": ["Users"],
        "summary": "Create new user",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "name",
            "description": "User Name",
            "required": true,
            "schema": {
              "type": "string",
              "example": "MAson MOunt"
            }
          },
          {
            "in": "body",
            "name": "phone",
            "description": "Phone Number of the user",
            "required": true,
            "schema": {
              "type": "string",
              "example": "9998877799"
            }
          },
          {
            "in": "body",
            "name": "email",
            "description": "Email address of the user",
            "required": true,
            "schema": {
              "type": "string",
              "example": "mount@gmail.com"
            }
          },
          {
            "in": "body",
            "name": "password",
            "description": "Password chosen by the user",
            "required": true,
            "schema": { "type": "string", "example": "asd!~@122" }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Operation",

            "schema": {
              "title": "Sample Response",
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User created successfully !"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/user/login": {
      "post": {
        "tags": ["Users"],
        "summary": "Login user to the system",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "email",
            "description": "Email address of the account",
            "required": true,
            "schema": {
              "type": "string",
              "example": "mount@gmail.com"
            }
          },
          {
            "in": "body",
            "name": "password",
            "description": "Password of the account",
            "required": true,
            "schema": { "type": "string", "example": "asd!~@122" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",

            "schema": {
              "title": "Sample Response",
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Mason Mount"
                },
                "userID": {
                  "type": "string",
                  "example": "5fb694c0d884419aa8c5d33a"
                },
                "email": {
                  "type": "string",
                  "example": "mount@gmail.com"
                },
                "token": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySUQiOiI1ZmI2OTRjMGQ4ODQ0MTlhYThs"
                },
                "expiresIn": {
                  "type": "number",
                  "example": 3000
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Incorrect Email or Password"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/user/all": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users to the system",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Operation",

            "schema": {
              "title": "Sample Response",
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Mason Mount"
                  },
                  "_id": {
                    "type": "string",
                    "example": "5fb694c0d884419aa8c5d33a"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/user/password-reset-request": {
      "post": {
        "tags": ["Users"],
        "summary": "Initiate password reset request",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "email",
            "description": "Email of the user for sending password reset link",
            "required": true,
            "schema": { "type": "string", "example": "mount@gmail.com" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",

            "schema": {
              "title": "Sample Response",
              "type": "object",

              "properties": {
                "message": {
                  "type": "string",
                  "example": "Password reset mail sent"
                },
                "resetEmail": {
                  "type": "string",
                  "example": "mount@gmail.com"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Request Body",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Email is required"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User with this email does not exist"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/user/valid-password-reset-token": {
      "post": {
        "tags": ["Users"],
        "summary": "Validates the password reset token",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "token",
            "description": "Passwordreset token sent with the password reset link",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySUQiOiI1ZmI2OTRjMGQ4ODQ0M"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",

            "schema": {
              "title": "Sample Response",
              "type": "object",

              "properties": {
                "validToken": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          },
          "400": {
            "description": "Invalid Request Body",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Token is required"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/api/user/reset-password": {
      "post": {
        "tags": ["Users"],
        "summary": "Resets password for the user",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "resettoken",
            "description": "Reset token sent with the password reset link",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySUQiOiI1ZmI2OTRjMGQ4ODQ0MTlhYThs"
            }
          },
          {
            "in": "body",
            "name": "newPassword",
            "description": "New Password",
            "required": true,
            "schema": { "type": "string", "example": "asd!~@122" }
          },
          {
            "in": "body",
            "name": "confirmedPassword",
            "description": "Confirmed Password",
            "required": true,
            "schema": { "type": "string", "example": "asd!~@122" }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Operation",

            "schema": {
              "title": "Sample Response",
              "type": "object",
              "properties": {
                "resetSuccess": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          },
          "400": {
            "description": "Invalid Request Body",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Some parameters are missing in the request body"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Account with this email does not exist"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "title": "Sample error response",
              "type": "object",
              "properties": {
                "error": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}